+++ Runtime/AR/ARUpdateOrder.cs
@@ -118,5 +118,10 @@ namespace UnityEngine.XR.ARFoundation
         /// </summary>
         public const int k_HumanBody = k_HumanBodyManager + 1;
 
+        /// <summary>
+        /// The <see cref="ARMeshManager"/>'s update order. Should come after
+        /// the <see cref="ARSession"/>.
+        /// </summary>
+        public const int k_MeshManager = k_Session + 1;
     }
 }